:umbrella: Initial docker-compose based CI! (#43)
We have Travis now! :tada:

The tests check that Odoo 8-10 base addon can be installed and that its tests succeed. They also test some facilities included in the image, such as the *.d folders.

There is also a test that clones the main scaffolding and runs its supported environments.

Fixes #6.