[DCK] Indent .yml files with 2 spaces
It turns out YAML gets much better indented with 2 spaces, but there are a bunch of 4-space files in the repo already.

To make it fine, we indent .yml (not .yaml, which are the one we use here). This way, for future migrations to docker-compose file format v3+, we can change the suffix and avoid merge conflicts, while keeping 2 spaces for new files.