[DCK] Remove inbox containers, use internal networks
The inbox proxies were a trick to isolate POP3/IMAP (or any other TCP) connections when a database got restored into a non-production environment.

Since Docker 1.10, internal networks are natively supported, which means that any in/out communication outside the container stack gets automatically blocked.

From now on, you can just configure incoming mail as usual in your Odoo instances with the safety that those incoming mails will not be read outside the production environment.

If you still need outside world communication for devel or test environments, you can enable it by switching the internal boolean, or by adding a public network with public access and attaching it to the containers that need it (usually just Odoo).

Note that for this to have a real effect, the `inverseproxy_shared` network must be internal too in test environments.

Fixes #98