Easier usage of ptvsd debugger
The command you need to use to have ptvsd working and correctly integrated is a little bit hard to do manually, and more with the proxies and all stuff integrated in the scaffolding.

Now, just set `PTVSD_ENABLE=1` when booting the container to autoprefix the command with ptvsd stuff.

Instructions on how to use new feature from https://github.com/Tecnativa/doodba-scaffolding/pull/4 are added too.