Make ptvsd great again
Without this patch, we were stick in an old ptvsd version, because it was the only one that was actually working, as explained in https://github.com/Microsoft/vscode-python/issues/133#issuecomment-353757512.

However, today it's the contrary: version 3.0.0 is no longer supported by Visual Studio Code, so we need to update it.

The new version has a different CLI, so it's adapted here too (BTW it was broken before too).

Usage instructions updated too.