Drop support for `$AUTO_REQUIREMENTS`
This feature was introduced in #71, but time has proven it buggy and not much useful.

This feature has a chicken-and-egg problem in the devel environment: code doesn't exist at build time, but it is expected to be able to find the `requirements.txt` files inside it.

It confuses the developer and creates divergencies between devel and test/prod environments (where that problem doesn't exist).

Also it's breaking #174, which is a more important and actually used feature.

The best fix is to drop support for this. If you really want to have requirements from remotes, you can add this to your `dependencies/pip.txt` file:

```
# Assuming you are installing all from server-tools
-r https://raw.githubusercontent.com/OCA/server-tools/11.0/requirements.txt
```

It's a better choice because:

- It's basic official pip.
- It's explicit.
- It works fine across environments.