Update python in v11 image (switch base)
The outdated and buggy Python 3.5 version that Debian 9 ships was triggering werid behaviors in some production instances.

I switch here the base to Docker Hub's `python:3.5-stretch`, which shares the same underlying OS, but uses latest Python 3.5 (3.5.6 as of today) and gets updates faster.

The main upgrade problem would be that projects adding `python-dev` to their `apt_build.txt` files should remove it now. Not a big problem nor fix, so it seems legit for the stable release.