Use Python 2+3 compatible code
I was using the `text=True` argument, only in PY3+, for `subprocess.check_call`, but it turns out that argument only exists in PY3.7+, and this code is executed also by PY2.7 and PY3.5, depending on the Odoo image version.

Instead, I just use now `universal_newlines=True`, which is the backwards-compatible equivalent to `text=True`. We'll change that when that backwards compatibility is removed and/or we only support PY3.7+ (a.k.a. only Odoo 12.0+).

Also, when using Python 2, the `closefd` argument for `os.fdopen` is not available, so we need to use a lower-level system to write bytes to STDOUT. This system is used only for Python 2, and marked to be removed when Odoo 10 support is dropped.