Add default fonts control
In #181 and #188 we added Liberation fonts as the default ones for reports. However this made it a little bit harder to choose a different default font (default fonts are the ones that the system interprets for `monospace`, `sans-serif` and `serif`).

From now on, if you want to override the default fonts, you have to add the package in `dependencies/apt.txt` or a `build.d` script, and add one or more of these new build arguments: `FONT_{MONO,SANS,SERIF}`. This way you make sure to override the default fonts. Before this, if you only added the package, it only made those fonts available, *not default*.