Fix wrong location of odoo script target
The odoo script must always point to the right executable in whatever odoo version you're using.

That's how it works for test and prod environments, but 89360140728b2e09c97cbd5bb7daae6ff581fbd0 introduced a regression on this matter for devel environment.

The main problem is that it was trying to be "smarter" by checking script existence, while at build time, in devel, those scripts still do not exist, so it always resolved to latest version (which uses `odoo-bin`) and broke older versions (those that use `openerp-server` or `odoo.py`).

I made it dumber-but-working again, and added a little comment to avoid future me to try to "fix" this again.