Add --enterprise to addons command
It's common that you want to test enterprise addons in a separate fashion from the rest of addons, same as core, extra or private addons.

To support that, from now on, a repo called `enterprise` will be excluded when you run `addons` with `--extra` flag. Just add `--enterprise` explicitly if you need it.

TT20969