Change no addons found exit code to 4
It turns out that an arguments syntax error exits with 0x2, which is the same exit code I chose to indicate there are no addons. This way, there's no way to know if there was a real error or there were simply no addons found.

So I change the exit code for addons not found to 0x4.

https://github.com/Tecnativa/doodba-qa/pull/17 won't be able to complete without this patch.

TT21246