Update Python 3 version in v11 and v12
Python 3.5 is flawed. It affects Odoo badly as shown in https://github.com/odoo/odoo/issues/20158. Odoo claims to have included a workaround for that in their code, but that simply doesn't work according to real world tests.

The only fix is to update Python version, so here's the update.

@Tecnativa TT21841