Fix CI
It seems the tested module `web_debranding` was removed from it-projects-llc/misc-addons repo (which is renamed now, BTW), so I hack the `ODOO_VERSION` env variable to use it as a commit which contains that module.

Also I'm pinning eslint<7 for v12, or it uses unsupported syntax and fails the test.