[FIX] autoaggrate: do not chown symlinks
when chowning symlinks, os.chown modifies the target instead of the symlink
as the target (if under SRC_DIR) should be chowned separately it's ok to skip
chowning symlinks. this prevents unclear error messages where it looks like the
symlink does not exist (when in fact it's the target file) and gives the same
result as if we are chowning targets via symlinks twice.