[FIX] 16/17: Proper ~root/.ssh pre-removal
The symlink to /opt/odoo/custom/ssh is still being created as sub-directory
(/root/.ssh/ssh), getting an error when trying to build the Doodba image.

Docker uses '/bin/sh' as default shell and '[[ ]]' is a specific comparison
construct of Bash.

Removing the brackets, now the symlinks to /opt/odoo/custom/ssh are being
successfully created.